+++ AAEmu.Game/Models/Account/AccountDetails.cs
@@ -4,7 +4,7 @@ namespace AAEmu.Game.Models.Account;
 
 public struct AccountDetails
 {
-    public int AccountId { get; set; }
+    public ulong AccountId { get; set; }
     public int AccessLevel { get; set; }
     public short Labor { get; set; }
     public int Credits { get; set; }
